mark objects no longer hold arguments in their properties
text("smthng") will return to {cls: Text, args: ["smthng"]}
it create Text mark with ...args and store it in controller's marks
(permaMarks for now). a mark may or may not store some args
(eg. Text doesn't). the refresh method also recieves the same arguments.
so if it's not needed to have it, there's no sense in storing it. just
update with the fresh one.