remove init() do in constructor instead
init() returns to the string that represents the html. it was only
needed in .grabFirstChil() and sync(). it only can be called once. it
doesn't just gives you the string with marked elements, as it marks
it also saves data. so calling will cause duplicates. also may not work
because you'll assign one string. so the previous marked elements won't
exist.

it should be called at some point. constructor looks like a good
place for it. also makes sense - new Vany should have a vResult. you'll
give it in argument. so now it can make the html string (which was done
with init). as this won't be modified ever again we do it right away.

the best thing in this way is that every step is clear.

let vanyl = new Vanyl(vResult) // creates the string in .html
we can assign this string to any element
also assign .topElement so Vanyl knows where to query
vanyl.process() // grab the marked elements for datas
vanyl.update() // do the update for datas