fix indexes for list. break untouched elements
when 1 2 3 goes 1 2 3 4 it was adding 4 and that's it
but if it goes like 1 2 4 3 it can't. not even 1 3 2.

the way I choose is to use fragments. add it all to a fragment and then
append all.

appending moves element from its current location to here. but cause the
css animation go again, as dom location changes.

maybe I wont't change it.