Code splitting; sharp decision for one element
Vanyls can only have one top element. so we cant sync a v that's
multiple html tags. have to be wrapped in one top tag.

it's the most vanilla approach - create an element and modify it.

also now we can new Vanyl(v`<b></b>`) and have a full Vanyl.

it's not how vanyl should be used. function approach is the preffered
way. but while developing, I used .grabFirstChild in a few places.
it's not so hard, I mean whole code is 150 lines. now my mind is
also pieceful knowing it will only be one element. having vResult is
enogh for most now.