list rendering moved to start; but problem
first it will place elements of list. but properties will be unchanged.
if it's positioned by data, then there will be animations correct.

first replace the elements in the correct positions. and then assign the
properties. this will make them to go to the positions from their olds.

but one problem - we need to use setTimeout. it doesn't runs animations
after appending them. maybe there's no chance for document to render
itself and then assigning property runs. so moving the element and
assigning the properties kinda goes in the "same cycle"?