Keep and Super Properties
Vanyl will append in the order in the given list. if the element isn't
supposed to be appended (like only update the css then position it
accordingly) then those list elements should be a keep super property to
be truthy

super key term just says that the first argument for the template should
be a property in the top element's dynamic property object.

v`<div ${{keep:true}}>${"that's not super property"}</div>`

for now, there's only 2 super prop. key; keep;

key - is used for element to be stored and inserted again in the future
updates
keep - don't re-append the elements on update

keyless lists will be implemented too. their elements will be created
each time. not efficient, not even useful for inputs. but a quick way to
show list of things.