remove ref; use do to assign
ref was just a function.
in update, the .element prop was setting.
and the function was returning that prop.
to use the ref, you should call it.
it either returns null or the element.

well, why not assign the actual variable?
use do: prop to assign it manually.
do is called with the element as the argument.
so you can use like { do: it => myvar = it }