Add opts(opts) to assigning props before construct
add class VanylEmbedded to use with the components.
component won't inherit from Vanyl. it inherits from HTMLElement. but we
need to have some way to update it with vanyl. the most basic way -
have a vanyl per component.
VanylEmbedded was to set these vanyls root as the component.

we couldn't set root before constructor fully called.
yes, getRoot method can be used, but we want to pass an element that's
not in that object. so should be assigned. calling opts in the top of
the constructor was the cleanest way. but it takes up 1 space in
arguments list. and that's the secon place for this time